blogicum/requirements.txt
asgiref==3.8.1
astroid==2.15.8
attrs==24.2.0
beautifulsoup4==4.12.3
Django==5.1.1
django-bootstrap5==24.3
django-debug-toolbar==5.2.0
Faker==12.0.1
flake8==6.1.0
flake8-django==1.4
flake8-docstrings==1.7.0
iniconfig==2.0.0
isort==6.0.1
lazy-object-proxy==1.11.0
mccabe==0.7.0
mixer==7.2.2
packaging==24.2
pep8-naming==0.14.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
py==1.11.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.1.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
sqlparse==0.5.2
tomli==2.1.0
wrapt==1.17.2
yapf==0.43.0
